Article에 member정보 저장

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -10,6 +10,8 @@ import com.example.demo.service.ArticleService;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 	
@@ -21,12 +23,15 @@ public class UsrArticleController {
 	
 	@GetMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
-		articleService.writeArticle(title, body);
+	public ResultData<Article> doWrite(HttpSession session, String title, String body) {
+		
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		articleService.writeArticle(loginedMemberId, title, body);
 		
 		int id = articleService.getLastInsertId();
 		
-		return ResultData.from("S-1", String.format("%d번 게시물을 작성했습니다.", id), articleService.getArticleById(id));
+		return ResultData.from("S-1", String.format("%d번 게시물을 작성했습니다.", id), articleService.forPrintArticle(id));
 	}
 	
 	@GetMapping("/usr/article/showList")
@@ -43,7 +48,7 @@ public class UsrArticleController {
 	@GetMapping("/usr/article/showDetail")
 	@ResponseBody
 	public ResultData<Article> showDetail(int id) {
-		Article foundArticle = articleService.getArticleById(id);
+		Article foundArticle = articleService.forPrintArticle(id);
 		
 		if (foundArticle == null) {
 			return ResultData.from("F-1", String.format("%d번 게시물은 존재하지 않습니다.", id));
@@ -54,11 +59,17 @@ public class UsrArticleController {
 	
 	@GetMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession session, int id, String title, String body) {
 		Article foundArticle = articleService.getArticleById(id);
 		
 		if (foundArticle == null) {
 			return ResultData.from("F-1", String.format("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		if (foundArticle.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다.");
 		}
 		
 		articleService.modifyArticle(id, title, body);
@@ -68,11 +79,17 @@ public class UsrArticleController {
 	
 	@GetMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
 		Article foundArticle = articleService.getArticleById(id);
 		
 		if (foundArticle == null) {
 			return ResultData.from("F-1", String.format("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		int loginedMemberId = (int) session.getAttribute("loginedMemberId");
+		
+		if (foundArticle.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
